Extract compatibility test helper functions into helpers.ts module
Remove the `runEslint` helper function for the compatibility tests from
the test file and move it into a helpers.ts module. This mirrors the
unit tests and cleans up the test file.
